# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp AI Business Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /faqs:
    get:
      operationId: listFaqs
      tags: [faqs]
      summary: List all FAQs
      parameters:
        - $ref: "#/components/parameters/BusinessIdParam"
      responses:
        "200":
          description: FAQ list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faq"
    post:
      operationId: createFaq
      tags: [faqs]
      summary: Create a new FAQ
      parameters:
        - $ref: "#/components/parameters/BusinessIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFaqBody"
      responses:
        "201":
          description: Created FAQ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
  /faqs/{id}:
    patch:
      operationId: updateFaq
      tags: [faqs]
      summary: Update a FAQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/BusinessIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFaqBody"
      responses:
        "200":
          description: Updated FAQ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
        "404":
          description: Not found
    delete:
      operationId: deleteFaq
      tags: [faqs]
      summary: Delete a FAQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/BusinessIdParam"
      responses:
        "204":
          description: Deleted

  # Canned Responses
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      parameters:
        - $ref: "#/components/parameters/BusinessIdParam"
      responses:
        "200":
          description: Service list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      parameters:
        - $ref: "#/components/parameters/BusinessIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/BusinessIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceBody"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/BusinessIdParam"
      responses:
        "204":
          description: Deleted

  # Bookings
...
  /webhook:
    get:
      operationId: verifyWebhook
      tags: [webhook]
      summary: WhatsApp webhook verification
      parameters:
        - name: hub.mode
          in: query
          schema:
            type: string
        - name: hub.verify_token
          in: query
          schema:
            type: string
        - name: hub.challenge
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Verification challenge
        "403":
          description: Forbidden
    post:
      operationId: receiveWebhook
      tags: [webhook]
      summary: Receive WhatsApp messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK

  # Dashboard
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get business settings
      parameters:
        - $ref: "#/components/parameters/BusinessIdParam"
      responses:
        "200":
          description: Business settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update business settings
      parameters:
        - $ref: "#/components/parameters/BusinessIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessSettings"

  # Simulate
...
components:
  parameters:
    BusinessIdParam:
      name: businessId
      in: query
      description: Business ID to scope the request to. Defaults to the first business if not provided.
      schema:
        type: integer
  schemas:
...
    Faq:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        answer:
          type: string
        keywords:
          type: array
          items:
            type: string
        hitCount:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - question
        - answer
        - keywords
        - hitCount
        - active
        - createdAt
        - updatedAt
    CreateFaqBody:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        keywords:
          type: array
          items:
            type: string
        active:
          type: boolean
          default: true
      required:
        - question
        - answer
        - keywords
    UpdateFaqBody:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        keywords:
          type: array
          items:
            type: string
        active:
          type: boolean
    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        duration:
          type: ["integer", "null"]
        category:
          type: ["string", "null"]
        active:
          type: boolean
        keywords:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - price
        - duration
        - category
        - active
        - keywords
        - createdAt
        - updatedAt
    CreateServiceBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        duration:
          type: integer
        category:
          type: string
        keywords:
          type: array
          items:
            type: string
        active:
          type: boolean
          default: true
      required:
        - name
    UpdateServiceBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        duration:
          type: integer
        category:
          type: string
        keywords:
          type: array
          items:
            type: string
        active:
          type: boolean
...
    BusinessSettings:
      type: object
      properties:
        id:
          type: integer
        businessName:
          type: string
        businessType:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        openingHours:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        currency:
          type: string
        customAiPrompt:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - businessName
        - businessType
        - phone
        - email
        - address
        - website
        - openingHours
        - description
        - currency
        - customAiPrompt
        - updatedAt
    UpdateSettingsBody:
      type: object
      properties:
        businessName:
          type: string
        businessType:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        website:
          type: string
        openingHours:
          type: string
        description:
          type: string
        currency:
          type: string
        customAiPrompt:
          type: string
...